addressed pr comments
